Compare output files by relative distance error
Previously, outputs were compared by file contents.

Now, the relative error of launch-landing distance
between the output and the reference has to stay
below a set maximum.

The sensitivity to a change in the inputs is high:
changing the molecular mass of helium by 500 ppb
caused the program to fail tests.